Adding cpp palindrome

+++ 16-palindrome/palindrome.cpp
@@ -7,7 +7,7 @@ bool isPalindrome(string text) {
   int i = text.length() - 1;
 
   for(auto letter : text) {
-    if(i <= text.length() / 2) {
+    if(i <= text.length() / (usize) 2) {
       cout << letter << " " << i  << " " << text[i] << letter == text[i] << endl;
     }
